requirements.txt
python-libsbml
nose
pandas
numpy
urllib3
pip>20
matplotlib
coverage